reorganized project files
